pyproject.toml
[tool.poetry]
name = "custom_playbooks"
version = "0.1.2"
description = ""
authors = ["Langleu <lars.lange@camunda.com>"]

[tool.poetry.dependencies]
python = "^3.8"
hikaru = "^0.11.0b0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
robusta-cli = "^0.9.12"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
